Offers flow change for Update Subscription case
Offers flow now automatically removes skus in skulist if they are
duplicates of oldSku. It also redirects to payment when it receives
an old sku and a sku list with exactly one item that is different from
the old sku. If the item is the same as the old sku, prints error to
console instead of opening offers flow.